Planet Placer as submitted to show general knowledge of Java
This game was created through the effort of a team of 3, myself
included, for my final project in java 2. The main objective of the game
is to place planets onto the screen which will revolve around the sun
depending on their starting distance from the sun and the initial speed
given to the planet. Along with these attributes, the user is able to
change the planet they are placing and it's size via sliders in the side
bar. The game comes on an end when two planets collide. Due to time
constraints we were unable to add an aesthetic introduction and game
over screen. In this project I was responsible for creating the main GUI
within the GameView and GameSpace classes, which had included the
SideBarSubLabel and SideBarSubPanel classes in order to clean up code
within the gameView class. In addition, I was responsible for the
implementation of the Planet class in order to ease the process of
creating new planets and storing them within a data structure. This
class had included a few static variables that would be consistent
options for all instances of the class. Along with creation of a planet
instance, this class was responsible for handling the movement of the
planet class. I broke this process up by functions of the different
variables that would need to be updated, gravity, velocity, and
position. For these functions I used Newton's laws of gravitation to
change the acceleration of the planet assuming the only force acting on
it is the force of gravity from the sun. I had hoped to research patched
conics in order to add a more realistic movement from gravity where the
other planets orbiting the sun would also affect the motion.